api/requirements.txt
connexion == 1.1.15
Werkzeug==0.16.1
python_dateutil == 2.6.0
setuptools >= 21.0.0
openstacksdk
python-crontab
python-openstackclient
